Ammo Powerup Added
To support the new ammunition system, an ammo powerup has been added to reload ammo.  The tripleshot was initially designed to unaffected by the ammunition system but it was not clear to the player that this was the case, so UI elements have been added to indicate to the player that the TripleShot powerup also provides a "Free Fire" mode.  Timer also provided so player does not waste ammo when TripleShot expires.