Bugfix - Enemy Laser
Added null check for player in enemy script to stop the enemy firing while the player is dead and to eliminate null reference exception.  I know there must be a better way to do this but I've yet to figure out how to communicate with an cloned prefab from a different script.