Proper bugfixes to replace previous temporary bugfix
Previous bugfix was temporary.  Using nullchecking to change program behavior is poor methodology for a finished product.  Figured out how to use an array and a for loop to order all remaining enemies on screen at time of player death to stop firing.  Also discovered and fixed a bug in the SpawnManager that was causing 1 enemy to spawn after the player's death before shutting down.