Major Bug Fixes and Redesign
No features added.  Englarged playfield and reduced scale to make more playspace for game.  Rebalanced mechanics to compensate.  Rewrote weapon firing and cooldown systems to be much less convoluted and efficient.   Major weapons bugfixes.  All wave laser types completely reworked to be stable.  I discovered the hard way that my original implementation was not actually something Unity was supposed to be capable of doing so it would work sometimes, and sometimes not.  I tried several different ways of fixing the problem before discovering that method wasn't possible and having to scrap it entirely.  All the wave lasers seem to be working perfectly in every game now.