Negative Powerup & Progress notes
Implemented a Powerdown.  Avoid the skull and crossbones or your ammo will be stolen and your speed cut in half for the duration.  Who can say for how long, as there is intentionally no feedback to the player on how long the effect will last.

The next objective is "Introduce New Enemy Type" but that was done in the first commit to this branch d538ea182c1021dcb85b1ff9b04042ed5ca7153a.  After that is "Balanced Spawning" which is also an already completed objective.  I created a balanced spawning system for powerups as part of  commit 5c74464a75e80ca85cb8cad59e026c37df295ca2 due to the demand of the objective that the wave laser powerup be a rare spawn.  I created the balanced enemy spawn system as part of the wave generator I built in commit 41a67980f3f809a48a28dd6141127f7223387660.